Move example bottom sheet to its own package, remove Foo naming
