+++ demo/src/main/java/co/zsmb/rainbowcake/demo/ui/example/ExampleFragment.kt
@@ -5,7 +5,7 @@ import android.view.View
 import co.zsmb.rainbowcake.base.RainbowCakeFragment
 import co.zsmb.rainbowcake.dagger.getViewModelFromFactory
 import co.zsmb.rainbowcake.demo.R
-import co.zsmb.rainbowcake.demo.ui.foo.FooBottomSheetFragment
+import co.zsmb.rainbowcake.demo.ui.bottomsheet.ExampleBottomSheetFragment
 import co.zsmb.rainbowcake.demo.ui.foo.FooFragment
 import co.zsmb.rainbowcake.demo.ui.koin.KoinFragment
 import co.zsmb.rainbowcake.demo.ui.sharedvmpager.SharedVMPagerFragment
@@ -42,7 +42,7 @@ class ExampleFragment : RainbowCakeFragment<ExampleViewState, ExampleViewModel>(
         }
 
         bottomSheetExampleDemoButton.setOnClickListener {
-            FooBottomSheetFragment().show(requireFragmentManager(), "FooBottomSheetFragment")
+            ExampleBottomSheetFragment().show(requireFragmentManager(), "FooBottomSheetFragment")
         }
     }
